Netmera - Json web servis dönüşüm 1. aşama

+++ booklet-ws/booklet-ws/src/main/java/com/fagose/booklet/model/BookLike.java
@@ -19,15 +19,15 @@ import javax.persistence.TemporalType;
 @Table(name = "BookLike", catalog = "booklet")
 public class BookLike implements java.io.Serializable {
 
-	private Integer bookLikeId;
-	private Integer bookLikerId;
-	private Integer bookId;
+	private Long bookLikeId;
+	private Long bookLikerId;
+	private Long bookId;
 	private Date bookLikeDate;
 
 	public BookLike() {
 	}
 
-	public BookLike(Integer bookLikerId, Integer bookId, Date bookLikeDate) {
+	public BookLike(Long bookLikerId, Long bookId, Date bookLikeDate) {
 		this.bookLikerId = bookLikerId;
 		this.bookId = bookId;
 		this.bookLikeDate = bookLikeDate;
@@ -36,29 +36,29 @@ public class BookLike implements java.io.Serializable {
 	@Id
 	@GeneratedValue(strategy = IDENTITY)
 	@Column(name = "BookLikeID", unique = true, nullable = false)
-	public Integer getBookLikeId() {
+	public Long getBookLikeId() {
 		return this.bookLikeId;
 	}
 
-	public void setBookLikeId(Integer bookLikeId) {
+	public void setBookLikeId(Long bookLikeId) {
 		this.bookLikeId = bookLikeId;
 	}
 
 	@Column(name = "BookLikerID")
-	public Integer getBookLikerId() {
+	public Long getBookLikerId() {
 		return this.bookLikerId;
 	}
 
-	public void setBookLikerId(Integer bookLikerId) {
+	public void setBookLikerId(Long bookLikerId) {
 		this.bookLikerId = bookLikerId;
 	}
 
 	@Column(name = "BookID")
-	public Integer getBookId() {
+	public Long getBookId() {
 		return this.bookId;
 	}
 
-	public void setBookId(Integer bookId) {
+	public void setBookId(Long bookId) {
 		this.bookId = bookId;
 	}
